extract createNewEpoch
