+++ lib/exclude.js
@@ -4,9 +4,74 @@
 
 const chunk = require('lodash.chunk')
 const pull = require('pull-stream')
+const { SecretKey } = require('ssb-private-group-keys')
 const clarify = require('clarify-error')
+const {
+  validator: {
+    group: { initEpoch: isInitEpoch },
+  },
+  keySchemes,
+} = require('private-group-spec')
 const Epochs = require('./epochs')
 
+function createNewEpoch(ssb, groupId, opts, cb) {
+  ssb.metafeeds.findOrCreate(function gotRoot(err, myRoot) {
+    // prettier-ignore
+    if (err) return cb(clarify(err, "Couldn't get own root when excluding members"))
+
+    const newSecret = new SecretKey()
+    const addInfo = { key: newSecret.toBuffer() }
+
+    ssb.box2.addGroupInfo(groupId, addInfo, (err) => {
+      // prettier-ignore
+      if (err) return cb(clarify(err, "Couldn't store new key when excluding members"))
+
+      const newKey = {
+        key: newSecret.toBuffer(),
+        scheme: keySchemes.private_group,
+      }
+      ssb.box2.pickGroupWriteKey(groupId, newKey, (err) => {
+        // prettier-ignore
+        if (err) return cb(clarify(err, "Couldn't switch to new key for writing when excluding members"))
+
+        const newEpochContent = {
+          type: 'group/init',
+          version: 'v2',
+          secret: newSecret.toString('base64'),
+          tangles: {
+            members: { root: null, previous: null },
+          },
+          recps: [groupId, myRoot.id],
+        }
+        const newTangleOpts = {
+          tangles: ['epoch'],
+          isValid: isInitEpoch,
+        }
+
+        ssb.tribes2.publish(newEpochContent, newTangleOpts, (err) => {
+          // prettier-ignore
+          if (err) return cb(clarify(err, "Couldn't post init msg on new epoch when excluding members"))
+
+          // prettier-ignore
+          if (opts?._reAddCrash) return cb(new Error('Intentional crash before re-adding members'))
+
+          reAddMembers(
+            ssb,
+            groupId,
+            { _reAddSkipMember: opts?._reAddSkipMember },
+            (err) => {
+              // prettier-ignore
+              if (err) return cb(clarify(err, "Couldn't re-add members when creating new epoch"))
+              cb()
+            }
+          )
+        })
+      })
+    })
+  })
+}
+
+/** idempotent */
 function reAddMembers(ssb, groupId, opts, cb) {
   const epochs = Epochs(ssb)
 
@@ -49,5 +114,6 @@ function reAddMembers(ssb, groupId, opts, cb) {
 }
 
 module.exports = {
+  createNewEpoch,
   reAddMembers,
 }
